backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13.0,<14
python-multipart==0.0.6

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml>=6.0,<7

# Data processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# Scientific analysis (available in code tool)
scipy>=1.11,<2
statsmodels>=0.14,<0.15
scikit-learn>=1.3,<2

# LLM providers
openai==1.10.0
anthropic==0.18.0
google-genai==0.2.2

# Optional: for vLLM compatibility (uses OpenAI client)
# vLLM is run separately, not installed as dependency
# Cron scheduling
croniter>=1.3

# HTTP Client
httpx==0.28.1

# Feishu / Lark gateway (optional — enabled via FEISHU_GATEWAY_ENABLED)
lark-oapi>=1.4.0  # Optional: Feishu gateway (only needed if FEISHU_GATEWAY_ENABLED=true)
